server/django_api/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
# Driver de MySQL
mysqlclient